Add rocketmail to freemails, implement reply IDs
Add a new script, run at beginning of page load on
http://cs.gumtree.com.au/replyts/screening.do* that checks for a message
ID in the URL and sets the value of the message ID field accordingly.
This allows direct links by message ID in ReplyTS.  Also add
rocketmail.com to the list of free email domains and fix the body ID
check in background.js so scripts will run only in Manage Ads.