Remove deletion countdown, replace with OK/cancel
Be smarter with our deletion confirmation: instead of counting down,
display a message directly below the select box with OK and cancel
buttons; as soon as an action is taken on another ad, confirm any other
pending deletions. Should also confirm pending deletions on unload, but
that appears to be more difficult.